core: z_media_preview: Do not add white space to small images when resizing
This commit changes the way Zotonic handles resizing images which are
resized beyond their original image dimensions. (when the 'upscale'
parameter is not used).

Previously, Zotonic would add a white area to the output image so that
the output image always had the requested size. However, this behaviour
turns out to be very annoying and can be emulated with CSS when needed.

This commit changes this behaviour so that for small images, no white
space is added but the image is left unresized.

(cherry picked from commit aab664cf6cc64faa7fb5866042b3ad2d3dcad00b)